completes qn 3 a,b

+++ computationalFinanceProject3/Mutils.h
@@ -15,7 +15,7 @@ public:
     static double StDev(long double arr[], int n);
     static double Corr(long double *x, long double *y, int size);
     static void WriteToCSV(long double *arr, int n, const string filename);
-    static double pnorm(double x);
+    static long double pnorm(double x);
     static double Cov(long double *x, long double *y, int size);
     static long double * MatrixMultiply(long double *arr, int size, double sqrtT);
     static long double * MatrixAddition(long double *arr, int size, long double num);
